feat: Add authentication, meeting, and interview modules with security configuration
Authentication & Security:
- Added AuthController and AuthService with JWT + Microsoft OAuth integration
- Implemented PKCE flow for Microsoft OAuth with token verification using RSA public keys
- Created DTOs for authentication (LoginRequest, LoginResponse, OAuthLoginRequest, RefreshRequest)
- Configured SecurityConfig with Spring Security, CSRF disabled for REST APIs
- Added global exception handling via GlobalExceptionHandler with Microsoft reauth detection

Interview Management:
- Added comprehensive interview functionality with InterviewController, InterviewService, and InterviewRepository
- Implemented role-based access control (RBAC) with @PreAuthorize annotations
- Created DTOs for interview operations (InterviewLogRequest, InterviewResponse, InterviewScheduleRequest, InterviewUpdateRequest)
- Added interview scheduling, logging, updating, and deletion capabilities

Core Models & Database:
- Created models for core entities (Committee, Interview, Subcommittee, User, UserSupervisor)
- Added Microsoft token storage fields to User model for persistent OAuth integration

Microsoft Integration & Utilities:
- Implemented GraphUserTokenProvider with automatic token refresh and reauth handling
- Added TeamsMeetingService for Microsoft Teams meeting creation via Graph API
- Created JwtUtil for JWT token generation and validation
- Added HierarchyService for organizational structure management

Testing:
- Added unit tests for HierarchyService and InterviewService

This commit introduces the foundational backend structure for authentication, meeting, and interview management, along with comprehensive security configurations and Microsoft Graph API integration.

+++ src/main/java/com/hertssu/meetings/service/MeetingService.java
@@ -12,12 +12,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import com.hertssu.meetings.model.Meeting;
-import com.hertssu.meetings.model.User;
 import com.hertssu.meetings.repository.MeetingRepository;
-import com.hertssu.meetings.repository.UserRepository;
+import com.hertssu.model.Meeting;
+import com.hertssu.model.User;
 
 import jakarta.persistence.EntityNotFoundException;
+import com.hertssu.user.UserRepository;
 
 /**
  *
@@ -33,7 +33,7 @@ public class MeetingService {
     }
 
     public Meeting createMeeting(Meeting meeting){
-        Integer creatorId = meeting.getCreator().getId();
+        Long creatorId = meeting.getCreator().getId();
 
         User creator = userRepository.findById(creatorId)
             .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + creatorId));
